Updated scripts

+++ scripts/shutdown.bat
@@ -4,7 +4,12 @@ setlocal
 
 set /p VERSION=<VERSION
 
-docker-compose down --volumes --remove-orphans
+if "%*"=="--production" (
+    docker-compose -f docker-compose-prod.yml down --volumes --remove-orphans
+) else (
+    docker-compose down --volumes --remove-orphans
+)
+
 docker container prune -f
 docker volume prune -f
 docker system prune -f
